// src/pages/Account_EditAccount/interface.tsx
import { I_Breadcrumbs } from "../../component/breadcrumbs/interface";
export type I_PageType = {
  breadcrumbs: I_Breadcrumbs;
  personal: string;
  account: string;
  requiredInput: string;
  surname: string;
  name: string;
  changePass: string;
  oldPasswordTitle: string;
  newPasswordTitle: string;
  requiredCharacters: string;
  passwordAgain: string;
  required: string;
  saveBtn: string;
  accountDelete: string;
  confirmDelete: string;
  deleteAccount: string;
};